Controllers for users and change the route to inclue api and version
